Add usePerils hook
